+++ src/KingmakerDiceRoller/ProductMetadata.cs
@@ -2,6 +2,6 @@ namespace KingmakerDiceRoller
 {
     internal static class ProductMetadata
     {
-        internal const string Version = "0.1.8";
+        internal const string Version = "0.1.9";
     }
 }